Added a few updated tests and fixed up how methods_exist works in order to deal with some bugs caused by switching to unions. On the bright side methods_exist runs alot faster now that it doesn't use methodswith.
Issues with methods_exist that led to this solution:
1) The initial problem was that we couldn't use methodswith cause any existing common functions using the interface type will get mixed with actual required method declarations. So we needed a way to distinguish between these 2 groups of methods by storing them some how.
2) Next tried to just make the interface methods part of the Interfaces module cause, allowing us to narrow the methodswith call to the Interfaces module, but we can’t add functions to a module namespace after the fact.
3) Finally, we build up a list of the method signatures and create a get_methods(::Type{<TheInterface>}) = return <methods> which we can then use from the methods_exist funciton to get our list. The rewrite of methods_exist to use this methods list returned from the InterfaceType specific get_methods allows us to avoid any calls to methodswith, which is slow.

NOTE: While debugging the above problem there is also a bug where methods using a typealias with and empty mutable union wouldn't update to support new types as we updated the typealias union. The solution for this for now is to automatically add a _<Interface> type to the union when we create the Interface. While this is ugly when looking at the method signatures the end user can't accidentally instantiate a _<Interface> and pass it in, so the behaviour still works as intended.